requirements.txt
# 区块链学习项目 - 依赖包
# 安装命令：pip install -r requirements.txt

web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
pytest>=7.4.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0
